requirements.txt
appdirs==1.4.4
attrs==20.1.0
black==20.8b1
cfgv==3.2.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
identify==1.4.29
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
mypy-extensions==0.4.3
nodeenv==1.5.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
pre-commit==2.7.1
psycopg2==2.8.5
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.0.1
PyYAML==5.3.1
regex==2020.9.27
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
virtualenv==20.0.31
Werkzeug==1.0.1
cryptography==2.8
boto3~=1.12.39
requests~=2.23.0
botocore~=1.15.39
